move ui related classes to ui package
