+++ app/src/androidTest/java/nl/tudelft/cs4160/trustchain_android/GuiEspressoTest/OverviewConnectionsActivityTest.java
@@ -6,7 +6,7 @@ import org.junit.Rule;
 import org.junit.Test;
 
 import nl.tudelft.cs4160.trustchain_android.R;
-import nl.tudelft.cs4160.trustchain_android.main.OverviewConnectionsActivity;
+import nl.tudelft.cs4160.trustchain_android.ui.main.OverviewConnectionsActivity;
 
 import static android.support.test.InstrumentationRegistry.getInstrumentation;
 import static android.support.test.espresso.Espresso.onView;
