+++ app/src/androidTest/java/nl/tudelft/cs4160/trustchain_android/main/ChainInfoTest.java
@@ -18,6 +18,7 @@ import org.junit.runner.RunWith;
 
 import nl.tudelft.cs4160.trustchain_android.R;
 import nl.tudelft.cs4160.trustchain_android.storage.sharedpreferences.UserNameStorage;
+import nl.tudelft.cs4160.trustchain_android.ui.userconfiguration.UserConfigurationActivity;
 
 import static android.support.test.InstrumentationRegistry.getInstrumentation;
 import static android.support.test.espresso.Espresso.onView;
