+++ app/src/androidTest/java/nl/tudelft/cs4160/trustchain_android/main/ViewMyChainTest.java
@@ -19,6 +19,7 @@ import org.junit.runner.RunWith;
 
 import nl.tudelft.cs4160.trustchain_android.R;
 import nl.tudelft.cs4160.trustchain_android.storage.sharedpreferences.UserNameStorage;
+import nl.tudelft.cs4160.trustchain_android.ui.userconfiguration.UserConfigurationActivity;
 
 import static android.support.test.InstrumentationRegistry.getInstrumentation;
 import static android.support.test.espresso.Espresso.onData;
