+++ app/src/main/AndroidManifest.xml
@@ -27,17 +27,17 @@
         android:theme="@style/AppTheme"
         tools:replace="android:allowBackup">
         <activity
-            android:name=".peersummary.PeerSummaryActivity"
+            android:name=".ui.peersummary.PeerSummaryActivity"
             android:configChanges="orientation|screenLayout|screenSize"
             android:windowSoftInputMode="stateUnchanged" />
-        <activity android:name=".chainExplorer.ChainExplorerActivity" />
-        <activity android:name=".main.ConnectionExplanationActivity" />
+        <activity android:name=".ui.chainexplorer.ChainExplorerActivity" />
+        <activity android:name=".ui.connectionexplanation.ConnectionExplanationActivity" />
         <activity
-            android:name=".main.OverviewConnectionsActivity"
+            android:name=".ui.main.OverviewConnectionsActivity"
             android:configChanges="orientation|screenLayout|screenSize"
             android:launchMode="singleInstance"/>
         <activity
-            android:name=".main.UserConfigurationActivity"
+            android:name=".ui.userconfiguration.UserConfigurationActivity"
             android:label="@string/app_name"
             android:noHistory="true"
             android:windowSoftInputMode="adjustResize">
@@ -52,12 +52,12 @@
         <activity android:name=".funds.qr.ExportWalletQRActivity" />
         <activity android:name=".funds.FundsActivity" />
         <activity
-            android:name=".chainExplorer.ChainExplorerInfoActivity"
+            android:name=".ui.chainexplorer.ChainExplorerInfoActivity"
             android:label="@string/title_activity_info_chain_explorer" />
         <activity
-            android:name=".main.ChangeBootstrapActivity"
+            android:name=".ui.changebootstrap.ChangeBootstrapActivity"
             android:windowSoftInputMode="adjustResize" />
-        <activity android:name=".inbox.InboxActivity" />
+        <activity android:name=".ui.inbox.InboxActivity" />
         <activity
             android:name=".offline.SendOfflineActivity"
             android:label="@string/title_activity_send_offline"></activity>
@@ -78,10 +78,10 @@
         <activity android:name=".passport.nfc.PassportConActivity" />
         <activity android:name=".offline.ScanQRActivity" />
         <activity
-            android:name=".stresstest.StressTestActivity"
+            android:name=".ui.stresstest.StressTestActivity"
             android:keepScreenOn="true" />
 
-        <service android:name=".service.NetworkConnectionService" />
-        <receiver android:name=".service.CancelNetworkConnectionReceiver" />
+        <service android:name=".network.NetworkConnectionService" />
+        <receiver android:name=".network.CancelNetworkConnectionReceiver" />
     </application>
 </manifest>
