--- app/src/main/java/nl/tudelft/cs4160/trustchain_android/main/DetailActivity.java
@@ -1,19 +1,0 @@
-package nl.tudelft.cs4160.trustchain_android.main;
-
-import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.Toolbar;
-
-import nl.tudelft.cs4160.trustchain_android.R;
-
-public class DetailActivity extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_detail);
-        Toolbar toolbar = findViewById(R.id.toolbar);
-        setSupportActionBar(toolbar);
-    }
-
-}
