+++ app/src/main/java/nl/tudelft/cs4160/trustchain_android/offline/SendOfflineActivity.java
@@ -32,7 +32,7 @@ import java.util.HashMap;
 
 import nl.tudelft.cs4160.trustchain_android.R;
 import nl.tudelft.cs4160.trustchain_android.funds.qr.QRGenerator;
-import nl.tudelft.cs4160.trustchain_android.main.OverviewConnectionsActivity;
+import nl.tudelft.cs4160.trustchain_android.ui.main.OverviewConnectionsActivity;
 import nl.tudelft.cs4160.trustchain_android.message.MessageProto;
 
 import static android.nfc.NdefRecord.createMime;
