+++ app/src/main/java/nl/tudelft/cs4160/trustchain_android/ui/chainexplorer/ChainExplorerAdapter.java
@@ -1,4 +1,4 @@
-package nl.tudelft.cs4160.trustchain_android.chainExplorer;
+package nl.tudelft.cs4160.trustchain_android.ui.chainexplorer;
 
 import android.app.Activity;
 import android.content.Context;
