+++ app/src/main/java/nl/tudelft/cs4160/trustchain_android/ui/changebootstrap/ChangeBootstrapActivity.java
@@ -1,4 +1,4 @@
-package nl.tudelft.cs4160.trustchain_android.main;
+package nl.tudelft.cs4160.trustchain_android.ui.changebootstrap;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -14,6 +14,7 @@ import java.net.InetAddress;
 
 import nl.tudelft.cs4160.trustchain_android.R;
 import nl.tudelft.cs4160.trustchain_android.storage.sharedpreferences.BootstrapIPStorage;
+import nl.tudelft.cs4160.trustchain_android.ui.main.OverviewConnectionsActivity;
 
 public class ChangeBootstrapActivity extends AppCompatActivity {
     private EditText bootstrapView;
