+++ app/src/main/java/nl/tudelft/cs4160/trustchain_android/ui/inbox/InboxActivity.java
@@ -1,4 +1,4 @@
-package nl.tudelft.cs4160.trustchain_android.inbox;
+package nl.tudelft.cs4160.trustchain_android.ui.inbox;
 
 import android.content.Context;
 import android.os.Bundle;
@@ -13,6 +13,7 @@ import java.util.ArrayList;
 import java.util.Collections;
 
 import nl.tudelft.cs4160.trustchain_android.R;
+import nl.tudelft.cs4160.trustchain_android.inbox.InboxItem;
 import nl.tudelft.cs4160.trustchain_android.peer.Peer;
 import nl.tudelft.cs4160.trustchain_android.storage.sharedpreferences.InboxItemStorage;
