+++ app/src/main/res/layout/activity_bluetooth.xml
@@ -27,7 +27,7 @@
             android:layout_width="match_parent"
             android:layout_height="match_parent"
             android:orientation="vertical"
-            tools:context="nl.tudelft.cs4160.trustchain_android.chainExplorer.ChainExplorerActivity">
+            tools:context="nl.tudelft.cs4160.trustchain_android.ui.chainexplorer.ChainExplorerActivity">
 
             <TextView
                 android:id="@+id/bluetooth_log"
