+++ app/src/main/res/layout/activity_chain_explorer_info.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="nl.tudelft.cs4160.trustchain_android.chainExplorer.ChainExplorerInfoActivity">
+    tools:context="nl.tudelft.cs4160.trustchain_android.ui.chainexplorer.ChainExplorerInfoActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
