+++ app/src/main/res/layout/activity_inbox.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="nl.tudelft.cs4160.trustchain_android.inbox.InboxActivity">
+    tools:context="nl.tudelft.cs4160.trustchain_android.ui.inbox.InboxActivity">
 
     <android.support.v7.widget.RecyclerView
         android:id="@+id/my_recycler_view"
