+++ app/src/main/res/layout/activity_peer_overview.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:id="@+id/myCoordinatorLayout"
-    tools:context="nl.tudelft.cs4160.trustchain_android.peersummary.PeerSummaryActivity">
+    tools:context=".ui.peersummary.PeerSummaryActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
