+++ app/src/main/res/layout/activity_stress_test.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".stresstest.StressTestActivity">
+    tools:context=".ui.stresstest.StressTestActivity">
 
     <LinearLayout
         android:orientation="vertical"
