--- app/src/main/res/layout/content_detail.xml
@@ -1,75 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context="nl.tudelft.cs4160.trustchain_android.main.DetailActivity"
-    tools:showIn="@layout/activity_detail">
-
-    <TableLayout
-        android:id="@+id/overviewButtons"
-        android:layout_width="match_parent"
-        android:layout_marginTop="20dp"
-        android:layout_height="wrap_content"
-        android:paddingLeft="@dimen/activity_horizontal_margin"
-        android:paddingRight="@dimen/activity_horizontal_margin">
-
-        <TableRow>
-
-            <TextView
-                android:layout_weight="1"
-                android:text="@string/local_ip" />
-
-            <TextView
-                android:id="@+id/local_ip_address_view"
-                android:layout_weight="1"
-                android:text="@string/ip" />
-
-        </TableRow>
-
-        <TableRow>
-
-            <TextView
-                android:layout_weight="1"
-                android:text="@string/wan_address" />
-
-            <TextView
-                android:id="@+id/wanvote"
-                android:layout_weight="1"
-                android:text="@string/unknown" />
-
-        </TableRow>
-
-        <TableRow>
-
-
-            <TextView
-                android:layout_weight="1"
-                android:text="@string/peer_id" />
-
-            <TextView
-                android:id="@+id/peer_id"
-                android:ellipsize="middle"
-                android:singleLine="true"
-                android:layout_weight="1"
-                android:text="" />
-
-        </TableRow>
-
-        <TableRow android:layout_marginBottom="10dp">
-
-            <TextView
-                android:layout_weight="1"
-                android:text="@string/connection" />
-
-            <TextView
-                android:id="@+id/connection_type"
-                android:layout_weight="1"
-                android:text="" />
-        </TableRow>
-
-    </TableLayout>
-
-</RelativeLayout>
