+++ app/src/main/res/layout/user_configuration.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="nl.tudelft.cs4160.trustchain_android.main.UserConfigurationActivity">
+    tools:context="nl.tudelft.cs4160.trustchain_android.ui.userconfiguration.UserConfigurationActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
